# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication/Authorization Service",
        "title": "Sentinel",
        "contact": {},
        "license": {
            "name": "AGPL-3.0 (With additional terms, see NOTICE file)",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify that user is logged-in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verifies user authentication",
                "operationId": "verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token in Token Bearer format",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdto.Payload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Create new access and refresh tokens and update current session info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes auth tokens",
                "operationId": "refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh Token (sent as HTTP-Only cookie in actual requests)",
                        "name": "X-Refresh-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Login endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login into the service",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "User credentials and audience",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestbody.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Logout endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke user session",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
        "/auth/oauth/introspect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "RFC 7662 (https://datatracker.ietf.org/doc/html/rfc7662). Valid token types are: access, refresh and activate.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "OAuth 2.0 Token Introspection",
                "operationId": "oauth-introspect",
                "parameters": [
                    {
                        "description": "OAuth2.0 token which must be introspected",
                        "name": "Token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestbody.Introspect"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Introspection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
        "/user/activate/{token}": {
            "get": {
                "description": "Activate user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activation"
                ],
                "summary": "Activate user",
                "operationId": "activate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requestbody.Auth": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "login": {
                    "type": "string",
                    "example": "admin@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "your-password"
                }
            }
        },
...
        "requestbody.Introspect": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJFZER..."
                },
                "type": {
                    "type": "string",
                    "example": "access"
                }
            }
        },
...
        "responsebody.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error"
                },
                "message": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "responsebody.Introspection": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "aud": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "urn:api:auth",
                        "urn:api:billing"
                    ]
                },
                "exp": {
                    "type": "integer",
                    "example": 1753707388
                },
                "iat": {
                    "type": "integer",
                    "example": 1753706788
                },
                "iss": {
                    "type": "string",
                    "example": "3c23ebbd-42af-47c6-9c50-7295b3ac3a62"
                },
                "jti": {
                    "type": "string",
                    "example": "ade1cdb0-309c-48c5-8251-c3f39ec0d606"
                },
                "scope": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read",
                        "write"
                    ]
                },
                "sub": {
                    "type": "string",
                    "example": "c9fcc8e3-f4f1-4b85-a65e-29bb889cbccb"
                }
            }
        },
...
        "responsebody.Token": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "expiresIn": {
                    "type": "integer",
                    "example": 600
                },
                "message": {
                    "type": "string",
                    "example": "hello"
                }
            }
        },
...
        "userdto.Payload": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "urn:api:auth",
                        "urn:api:billing",
                        "https://example.domain.com"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "d529a8d2-1eb4-4bce-82aa-e62095dbc653"
                },
                "login": {
                    "type": "string",
                    "example": "admin@mail.com"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user",
                        "moderator"
                    ]
                },
                "session-id": {
                    "type": "string",
                    "example": "35b92582-7694-4958-9751-1fef710cb94d"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token format: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRF_Cookie": {
            "description": "CSRF token in cookie. Required for state-changing requests, must match with token in X-CSRF-Token header",
            "type": "apiKey",
            "name": "_csrf",
            "in": "cookie"
        },
        "CSRF_Header": {
            "description": "CSRF token in request header. Required for state-changing requests, must match with token in cookie",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}